requirements-dev.txt
-r requirements.txt
pytest==6.2.2
pytest-cov==2.11.1
pytest-django==4.1.0
